requirements.txt
requests
feedparser
google-generativeai